requirements.txt
streamlit>=1.37.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
transformers>=4.41.0
lpips>=0.1.4
